+++ AGENTS.md
@@ -108,7 +108,7 @@ Syntax error in text   ← ⚠️ Mermaid 生成失敗
 
 ---
 
-## 🔧 Agent Check List
+## 🔧 Check List
 
 1. **流程順序**  
    - 讀取原始檔 → **題幹抽取** → 清洗 → 產出  
@@ -129,18 +129,4 @@ Syntax error in text   ← ⚠️ Mermaid 生成失敗
 5. **心智圖生成**  
    - Prompt 第一行必須是 `mindmap`，且輸出不得再包 ```mermaid。  
 
-6. **單元測試**  
-   - 為 Issue 1–4 各建立最小測例，驗證修正後僅輸出題幹。  
 
----
-
-## ⏩ 回報格式
-
-```text
-### Issue X
-Root-Cause：
-Fix Plan：
-Validation：
-```
-
-完成後彙總所有 Issue；如發現其他潛在問題，亦請以相同格式補充。
